Refactor extracting ClassResolverCache class
